requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
pydantic==1.10.9
python-dotenv==1.0.0